+++ ui/pages/Search/components/filters/LabeledSlider.jsx
